.htaccess: Add cache-control
This forces browsers to re-validate their cache. Previously, it was up to the browser to determine when to update the cache - it was provided `ETAG`s and `Last-Modified` in most cases, but no directive on how to handle them. This could lead to - depending on the browser and mainly its default settings - very long storage times where websites are only updated once in a blue moon. This is btw. to my knowledge not at all unified over all browsers, another reason to provide this information from the website. I know of no modern homepage not specifying some form of cache-control/expires rule, making it even more reasonable to have this rule also on at least our webspace.
Btw. this does not lead to significant data transfer, it in most cases only leads to `If-None-Match` and `If-Modified-Since`-requests that in most cases are answered with `304 Not Modified`. More information: https://developer.mozilla.org/docs/Web/HTTP/Caching